PromptMenu instantiation moved to constructor

+++ Headers/UserInterface/UserInterface.h
@@ -199,6 +199,6 @@ private:
 
 
     const char* debugMessage;
-    int debugWidth;
-    int debugHeight;
+    float debugWidth;
+    float debugHeight;
 };
